requirements.txt
flask
tensorflow
keras
pillow
openai
python-dotenv
flask_sqlalchemy